inference with tf lite quantized model

+++ edge/inference-service/inference_app.py
@@ -14,7 +14,7 @@ def perform_inference():
     }
     """
     data = request.get_json(force=True)
-    image = pickle.loads(data['image'].encode('latin-1'))
+    image = pickle.loads(data['image'].encode('latin-1')).astype(np.float32)
 
     found_something, detected_animals = infer(app.model, image)
